# sql: cleanup making copies of the eval context