This PR audits all processors to avoid now-redundant (as of #125715) copies of the `eval.Context` as well as performs some miscellaneous cleanup along the way.

Some microbenchmarks:
```
name                   old time/op    new time/op    delta
LastWriteWinsInsert-8    1.37ms ±13%    1.41ms ±16%    ~     (p=0.265 n=20+20)

name                   old alloc/op   new alloc/op   delta
LastWriteWinsInsert-8     253kB ± 2%     247kB ± 2%  -2.65%  (p=0.000 n=19+20)

name                   old allocs/op  new allocs/op  delta
LastWriteWinsInsert-8     1.67k ± 0%     1.67k ± 0%  -0.18%  (p=0.033 n=18+19)
```
```
FlowSetup/vectorize=true/distribute=true-8       144µs ± 3%     142µs ± 3%   -1.25%  (p=0.005 n=25+25)
FlowSetup/vectorize=true/distribute=false-8      139µs ± 6%     138µs ± 3%     ~     (p=0.260 n=24+22)
FlowSetup/vectorize=false/distribute=true-8      137µs ± 2%     138µs ± 4%     ~     (p=0.158 n=22+25)
FlowSetup/vectorize=false/distribute=false-8     132µs ± 2%     133µs ± 3%   +1.21%  (p=0.001 n=24+25)

name                                          old alloc/op   new alloc/op   delta
FlowSetup/vectorize=true/distribute=true-8      22.9kB ± 9%    18.9kB ± 5%  -17.18%  (p=0.000 n=24+21)
FlowSetup/vectorize=true/distribute=false-8     21.3kB ±10%    17.7kB ±12%  -17.04%  (p=0.000 n=22+22)
FlowSetup/vectorize=false/distribute=true-8     27.1kB ± 0%    25.3kB ± 0%   -6.62%  (p=0.000 n=20+20)
FlowSetup/vectorize=false/distribute=false-8    25.6kB ± 0%    23.8kB ± 0%   -6.99%  (p=0.000 n=20+20)

name                                          old allocs/op  new allocs/op  delta
FlowSetup/vectorize=true/distribute=true-8         234 ± 2%       231 ± 1%   -1.01%  (p=0.000 n=20+21)
FlowSetup/vectorize=true/distribute=false-8        219 ± 8%       217 ± 7%   -0.93%  (p=0.016 n=22+22)
FlowSetup/vectorize=false/distribute=true-8        229 ± 0%       228 ± 0%   -0.44%  (p=0.000 n=21+21)
FlowSetup/vectorize=false/distribute=false-8       212 ± 0%       211 ± 0%   -0.47%  (p=0.000 n=20+20)
```

Epic: CRDB-39063.