@mgartner 2024-06-25
*[`pkg/sql/execinfra/processorsbase.go` line 794 at r4](https://reviewable.io/reviews/cockroachdb/cockroach/126036#-O0B-dFL8oNuCC-lDnDI:-O0FG_r57pQkRfKOAySI:bow9cs6) ([raw file](https://github.com/cockroachdb/cockroach/blob/c67bd53a5c44109f3cfe8e52c9fc8c429782ffbb/pkg/sql/execinfra/processorsbase.go#L794)):*
<details><summary><i>Previously, yuzefovich (Yahor Yuzefovich) wrote…</i></summary><blockquote>

Yeah, that's the only mutation AFAIK. The problem is that we might have concurrency between processors running on the same node, so if two processors were to use the same `eval.Context` via the ExprHelper, we could have corruption / data races.
</blockquote></details>

Ahh makes sense. Thanks for explaining.


<!-- Sent from Reviewable.io -->